# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: NYL Labs API
host: 'localhost:10010'
basePath: /
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  '/users/{userId}':
    x-swagger-router-controller: userController
    get:
      description: Returns User Details to the caller
      operationId: getUser
      parameters:
        - name: userId
          in: path
          type: string
          required: true
          description: The id of the user
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/User'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      description: >-
        Delete The specified user when he doesn't buy a policy or tries to
        create multiple accounts
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          type: string
          required: true
          description: The id of the user
      responses:
        '204':
          description: No content returned on deletion
    patch:
      description: Update the User Details
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          type: string
          required: true
          description: The id of the user
        - name: user
          in: body
          description: User Update Details
          required: true
          schema:
            $ref: '#/definitions/UserDetails'
      responses:
        '200':
          description: Success message on updating User Details
          schema:
            $ref: '#/definitions/SuccessResponse'
        default:
          description: Error while creating recommendation input data
          schema:
            $ref: '#/definitions/ErrorResponse'
  /users:
    x-swagger-router-controller: userController
    post:
      description: Creates a new user from a third party app
      operationId: createUser
      parameters:
        - name: user
          in: body
          description: User
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '204':
          description: Message on user creation
          schema:
            $ref: '#/definitions/SuccessResponse'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
  User:
    type: object
    properties:
      address:
        type: object
        properties:
          add_line_1:
            type: string
          add_line_2:
            type: string
          city:
            type: string
          country:
            type: string
          state:
            type: string
          zip:
            type: string
        required:
          - add_line_1
          - add_line_2
          - city
          - country
          - state
          - zip
      date_of_birth:
        type: string
      dl_dtls:
        type: object
        properties:
          date_of_birth:
            type: string
          dl_number:
            type: string
          first_name:
            type: string
          last_name:
            type: string
          middle_name:
            type: string
        required:
          - date_of_birth
          - dl_number
          - first_name
          - last_name
          - middle_name
      email_id:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone:
        type: array
        items:
          type: object
          properties:
            number:
              type: string
            type:
              type: string
          required:
            - number
            - type
      ssn:
        type: string
      user_id:
        type: string
  UserDetails:
    type: object
    properties:
      address:
        type: object
        properties:
          add_line_1:
            type: string
          add_line_2:
            type: string
          city:
            type: string
          country:
            type: string
          state:
            type: string
          zip:
            type: string
        required:
          - add_line_1
          - add_line_2
          - city
          - country
          - state
          - zip
      date_of_birth:
        type: string
      dl_dtls:
        type: object
        properties:
          date_of_birth:
            type: string
          dl_number:
            type: string
          first_name:
            type: string
          last_name:
            type: string
          middle_name:
            type: string
        required:
          - date_of_birth
          - dl_number
          - first_name
          - last_name
          - middle_name
      email_id:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone:
        type: array
        items:
          type: object
          properties:
            number:
              type: string
            type:
              type: string
          required:
            - number
            - type
      ssn:
        type: string
...
  ErrorResponse:
    properties:
      message:
        type: string
      errorCode:
        type: number
      metadata:
        type: object
  SuccessResponse:
    required:
      - success
    properties:
      success:
        type: boolean